Alert message added while deleting something ot log out

+++ src/Components/ExistingActivity/Table/TableBody.jsx
@@ -5,6 +5,7 @@ import firestore from "../../../Firebase/Firestore";
 import UserContext from "../../../Context/UserContext/UserContext";
 import AllTrackContext from "../../../Context/AllTrackData/AllTrackContext";
 import Loading from "react-loading";
+import Swal from "sweetalert2";
 
 const TableBody = ({getData,collectionId,index}) => {
     const {Id,ConsumerName,ConsumerNumber,BillUnit,ConsumerMobileNumber,CreatedAt,MNREApplicationNumber , PVApplicationNumber} = getData || {};
@@ -17,6 +18,9 @@ const TableBody = ({getData,collectionId,index}) => {
 
     const [isLoading,setIsLoading] = useState(false);
 
+    // alert
+
+
 
     const handleLink = useCallback((event)=>{
         event.stopPropagation();
@@ -26,15 +30,33 @@ const TableBody = ({getData,collectionId,index}) => {
 
     const deleteHandle = (event)=>{
         event.stopPropagation();
-        setIsLoading(true)
-        firestore.deleteDocument(collectionId,Id).then(()=>{
-          firestore.getAllDocuments(user?.companyID+"TrackSolarData")
-          .then((data)=>{
-            setAllTrack(data)
-            setIsLoading(false)
-          })
-        })
+        // setIsLoading(true)
+        // 
 
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            firestore.deleteDocument(collectionId,Id).then(()=>{
+                firestore.getAllDocuments(user?.companyID+"TrackSolarData")
+                .then((data)=>{
+                  setAllTrack(data)
+                  setIsLoading(false)
+                })
+              })
+            Swal.fire({
+              title: "Deleted!",
+              text: "Your file has been deleted.",
+              icon: "success"
+            });
+          }
+        });
     }
